requirements.txt
openai
pyyaml
streamlit
PyPDF2
youtube_transcript_api
streamlit-option-menu